rearranging functions into their actual pages and logic folders
